==> notebooks/00_cleaning_and_wrangling.ipynb <==
import pandas as pd

metadata = pd.read_csv("../data/raw/metadata.csv")
metadata.head()

metadata.isna().sum()

metadata = metadata.drop_duplicates(subset=["canonical_composer", "canonical_title"], keep="last")
metadata.shape

metadata.drop(columns=["canonical_composer", "canonical_title", "year", "audio_filename"], inplace=True, errors="ignore")
metadata.head()

metadata.to_csv("../data/clean/metadata.csv", index=False)
